🚚 move components to their folders
