+++ src/components/layout/Layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Header from "./Header";
+import Header from "../header/Header";
 
 export default function Layout() {
   return (
